wolfy/pyproject.toml
[tool.poetry]
name = "wolfy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
wolfcrypt = "^5.6.6.post0"
click = "^8.1.7"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
